misc: Trimmed down the atlas77 header and renamed it
